+++ CHANGELOG.md
@@ -4,19 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### 2.0.0  - (Unreleased)
+### 2.0.0  - (2021-01-11)
 
 Most of the app has be rewritten in this update
 
 #### Added
 
 - Animated bottom bar
-- Meditation
+- Meditation timer
+- Bottom Sheets
+- Dialogs
 - Lots more languages
 
 #### Changed
 
-- New bottom sheets
 - Lists moved to main page
 - Lots of optimisations and fixes